sesh updates

+++ install.sh
@@ -2,5 +2,8 @@
 
 # patch for this issue running nvim in tmux in wezterm: https://github.com/wezterm/wezterm/issues/5007
 if [[ -f /etc/profile.d/wezterm.sh ]]; then
-  sudo cp ./before_wezterm.sh /etc/profile.d/before_wezterm.sh
+  # Check if the file does not already exist
+  if [ ! -f /etc/profile.d/before_wezterm.sh ]; then
+    sudo cp ./before_wezterm.sh /etc/profile.d/before_wezterm.sh
+  fi
 fi
